Check paths before deleting; better error handling

+++ test/install_test.ps1
@@ -24,7 +24,9 @@ $env:FLOSSBANK_INSTALL_TOKEN = "cf667c9381f7792bfa772025ff8ee93b89d9a757e6732e87
 
 # install the latest version at the default location
 $DefaultLocation = Join-Path $Home ".flossbank"
-Remove-Item -Force -Recurse $DefaultLocation
+if (!(Test-Path $DefaultLocation)) {
+  Remove-Item -Force -Recurse $DefaultLocation
+}
 $FLOSSBANK_INSTALL = ""
 .\install.ps1
 $Exe = Join-Path $DefaultLocation "bin" "flossbank"
@@ -32,7 +34,9 @@ Start-Process $Exe -Wait -NoNewWindow -PassThru
 
 # install to a custom location
 $CustomLocation = Join-Path $Home "flossbank-custom"
-Remove-Item -Force -Recurse $CustomLocation
+if (!(Test-Path $CustomLocation)) {
+  Remove-Item -Force -Recurse $CustomLocation
+}
 $FLOSSBANK_INSTALL = $CustomLocation
 .\install.ps1
 $Exe = Join-Path $CustomLocation "bin" "flossbank"
